refactor: extract toJSON to separate file
